requirements.txt
python-docx==0.8.11
flask==2.3.3
matplotlib==3.7.2
seaborn==0.12.2
numpy==1.24.3